Beginning of independent Game Infrastructure
With this type of infrastructure in mind, there will be no issues regarding the support of several games having an administration interface. You don't care about the details of the game. It's an abstraction layer on top of the game. The game will only manage it's stuff and administration starts a higher level. It will not be important how the game manages it, but that it can manage it.

Issue #8