Fix UI updating problem
The UI was not updating correctly. Updates on the datasource on another thread / task only appeared on an edit within the forms. With the new Context-property in the ViewModelBase which holds the SynchronizationContext, we are now able to post the updates onto the UI-Thread. But that was not all. Some PropertyChanged-Events were missing. They are now sent correctly. UI updates as expected.

Fixed Issue #14