Basic Game Layer Adapter for Battlefield 4
I created a basic adapter to the game layer for Battlefield 4. Through this adapter it is possible to send commands  and receive events. This Layer has to be introduced into the existing API-Controllers within the server. Afterwards the Controllers a independent from the games they support. The instance of the game is responsible for handling these commands and events.

Issue #8