Loading games implemented
All games supported by RaptoRCon must be available in the "Games"-subdirectory of the server. The IGame interface is the basic interface all game adapters have to implement. MEF is used for inspecting all assemblies within the stated directory for implementations of IGame. The IGamesContext can be used for retrieving the needed game(s).
For external use, the GamesController delivers all needed data.

Issue #8